int / int 는 int 인데 왜 double 로 나온다고 착각하는거지

+++ src/main/java/kr/teamcocoa/freefight/scoreboard/ScoreboardManager.java
@@ -93,7 +93,7 @@ public class ScoreboardManager {
         lines.add("K/D:");
         lines.add(getArrowMessage(decimalFormat.format((kills == 0 || deaths == 0)
                 ? 0
-                : kills / deaths)) + "    ");
+                : (double) kills / (double) deaths)) + "    ");
         lines.add("");
         lines.add(CurrentKitScoreboard.getInstance().getMessage(freeFightPlayer.getPlayer()) + ":");
         lines.add(getArrowMessage(Kits.getNameByEnum(freeFightPlayer.getCurrentKit())));
